Split architecture::arm::core::exceptions into separate files
